Updated src/Kernel/xxxKernel.cc to only instantiate kernels of type double. Updated the example code to use the new standard image and mask types. Modified examples/SConscript for the two examples that only use vw; one was listed as using fw, but it only ran if vw and boost were both specified; the other listed vw and ran, but I added boost "just to be sure".

+++ examples/analyticKernel.cc
@@ -13,7 +13,7 @@
 using namespace std;
 
 int main() {
-    typedef float pixelType;
+    typedef double pixelType;
     double sigmaX = 2.0;
     double sigmaY = 2.5;
     unsigned int kernelCols = 6;
